% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Internal Data Cleaning Function}
\usage{
clean_data(rawDT)
}
\arguments{
\item{rawDT}{[DESCRIPTION OF ARG NEEDED]}
}
\value{
A trimmed data.table
}
\description{
Internal Data Cleaning Function
}
