% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_connection.R
\name{conn_functions}
\alias{conn_functions}
\alias{build_conn_string}
\alias{conn_params}
\alias{get_conn_object}
\title{Internal Data Query Function}
\usage{
build_conn_string(cnList)

conn_params()

get_conn_object(cnString)
}
\arguments{
\item{cnList}{Internal argument}

\item{cnString}{Internal argument}
}
\description{
Internal Data Query Function
}
\section{Functions}{
\itemize{
\item \code{build_conn_string}: Internal function that builds the connection string using

\item \code{conn_params}: Internal function that decrypts and returns the connection params

\item \code{get_conn_object}: Internal function that returns the connection object
}}

