% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_functions.R
\name{get_data}
\alias{get_data}
\alias{getDataQTI}
\alias{getRawData}
\title{Internal Data Query Function}
\usage{
getDataQTI(hospID = NULL, cols = "*")

getRawData()
}
\arguments{
\item{hospID}{[DESCRIPTION OF ARGUMENT NEEDED]}

\item{cols}{columns to limit query to for minor efficiency gains}
}
\value{
[DESCRIPTION OF ARGUMENT NEEDED]
}
\description{
Internal Data Query Function
}
\section{Functions}{
\itemize{
\item \code{getDataQTI}: Returns raw data with minor filtering on the server for efficiency

\item \code{getRawData}: Returns raw data with no filtering or cleaning
}}

