% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_data.R
\name{get_raw}
\alias{get_raw}
\alias{get_conn}
\alias{get_raw}
\title{Internal Data Query Function}
\usage{
get_conn()

get_raw(hospID = NULL)
}
\arguments{
\item{hospID}{[DESCRIPTION OF ARGUMENT NEEDED]}
}
\value{
[DESCRIPTION OF ARGUMENT NEEDED]
}
\description{
Internal Data Query Function
}
\section{Functions}{
\itemize{
\item \code{get_conn}: [DESCRIPTION OF FUNCTION NEEDED]

\item \code{get_raw}: [DESCRIPTION OF FUNCTION NEEDED]
}}

