% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataQTI.R
\name{getDataQTI}
\alias{getDataQTI}
\title{[TITLE NEEDED]}
\usage{
getDataQTI(hospID = NULL)
}
\arguments{
\item{hospID}{[DESCRIPTION OF ARGUMENT NEEDED]}
}
\value{
[DESCRIPTION OF OUTPUT NEEDED]
}
\description{
[DESCRIPTION OF FUNCTION NEEDED]
}
