% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHospData.R
\name{getHospData}
\alias{getHospData}
\title{[TITLE NEEDED]}
\usage{
getHospData(SELECT = "*", WHERE = NULL)
}
\arguments{
\item{SELECT}{[DESCRIPTION OF ARGUMENT NEEDED]}

\item{WHERE}{[DESCRIPTION OF ARGUMENT NEEDED]}
}
\value{
[DESCRIPTION OF OUTPUT NEEDED]
}
\description{
[DESCRIPTION OF FUNCTION NEEDED]
}
