% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatPSM.R
\name{getStatPSM}
\alias{getStatPSM}
\title{Test Difference of Means with Balancing}
\usage{
getStatPSM(DT = NULL, treat_var = NULL, split_vars = NULL,
  covariates = NULL, outcome_var = NULL)
}
\arguments{
\item{DT}{[ARGUMENT DEFINITION NEEDED]}

\item{treat_var}{[ARGUMENT DEFINITION NEEDED]}

\item{split_vars}{[ARGUMENT DEFINITION NEEDED]}

\item{covariates}{[ARGUMENT DEFINITION NEEDED]}

\item{outcome_var}{[ARGUMENT DEFINITION NEEDED]}
}
\value{
[DESCRIPTION OF OUTPUT NEEDED]
}
\description{
[DESCRIPTION OF FUNCTION NEEDED]
}
