% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_data.R
\name{Match_Data}
\alias{Match_Data}
\alias{getMatchedData}
\title{Match Data Using PSM}
\usage{
getMatchedData(hospDT = NULL, stratby = NULL, covars = NULL,
  outcomes = NULL, match_var = "Modality")
}
\arguments{
\item{hospDT}{[ARGUMENT DEFINITION NEEDED]}

\item{stratby}{VARS TO STRATIFY THE DATA BY (PER REQUIREMENTS)}

\item{outcomes}{[ARGUMENT DEFINITION NEEDED]}

\item{covariates}{[ARGUMENT DEFINITION NEEDED]}
}
\description{
Match Data Using PSM
}
\section{Functions}{
\itemize{
\item \code{getMatchedData}: Takes in data from getDataQTI and returns a list of data.tables
with each list element being a PSM matched dataset containing
}}

