% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_stat.R
\name{morph_stat}
\alias{morph_stat}
\title{Patient Cohort Analysis}
\usage{
morph_stat(DAT = NULL, GRP_VARS = c("PRIMARY_PROCEDURE", "BENIGN_MALIGNANT",
  "PATIENT_TYPE"), TREATMENT = "MODALITY", OUTCOME, min_n = 10)
}
\arguments{
\item{DAT}{data.table}

\item{GRP_VARS}{Variables use to stratify observations into groups}

\item{TREATMENT}{Name of the column that represents assignment of treatment or control group}

\item{OUTCOME}{The name of the column that represents the outcome}

\item{min_n}{The minimum number of patients required to have a valid group (default = 25)}
}
\value{
A data.table with the desired statistics
}
\description{
Compute basic statistics as well as a paired T-Test on an outcome measure (\code{outcome}) after
stratification of the data using the provided covariates (\code{covars}) and
balancing cohorts of patients based on treatment modality (\code{treatment}).
}
\details{
Note: data provided in the \code{DT} argument should have carefully considered outliers and potential to influence.
Additionally, this function expects two groups. Therefore the treatment column should be reduced to the desired
two group values.
}
\examples{
# EXAMPLES PENDING
}
