% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm_stats.R
\name{PSM_Analysis}
\alias{PSM_Analysis}
\alias{get_stats}
\alias{psm_data}
\title{Test Difference of Means with Balancing}
\usage{
get_stats(DT, outcome_vars)

psm_data(DT = NULL, strat_vars = NULL, covariates = NULL,
  outcome_vars = NULL)
}
\arguments{
\item{DT}{[ARGUMENT DEFINITION NEEDED]}

\item{outcome_vars}{[ARGUMENT DEFINITION NEEDED]}

\item{strat_vars}{[ARGUMENT DEFINITION NEEDED]}

\item{covariates}{[ARGUMENT DEFINITION NEEDED]}
}
\value{
[DESCRIPTION OF OUTPUT NEEDED]
}
\description{
[DESCRIPTION OF FUNCTION NEEDED]
}
\section{Functions}{
\itemize{
\item \code{get_stats}: Function that takes the outcome of \code{getDataPSM} and 
runs statistical summary functions and performs a paired t-test across the psm matched
groups

\item \code{psm_data}: Takes in data from getDataQTI and returns a list of data.tables
with each list element being a PSM matched dataset containing
}}

