% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm_functions.R
\name{PSM_Functions}
\alias{PSM_Functions}
\alias{caller_env}
\alias{runPSM}
\alias{get_result_psm}
\alias{getMatchedData}
\alias{getStats}
\title{Run Full Analysis with PSM Matching}
\usage{
caller_env(n = 1)

runPSM(hospID = NULL)

get_result_psm(hDT, idcols, strat, covars, outcomes)

getMatchedData(DT = NULL, strat_vars = NULL, covariates = NULL,
  outcome_vars = NULL)

getStats(DT, outcome_vars)
}
\arguments{
\item{n}{internal argument for helper \code{caller_env}}

\item{hospID}{A numeric variable representing the hospital id}

\item{hDT}{Hospital data}

\item{idcols}{Columns to keep as IDs}

\item{strat}{Columns to stratify the hospital data before PSM}

\item{covars}{Covariates to use to calculate the propensity score to match on}

\item{outcomes}{Outcome column variables to run statistics for across groups matched with PSM}

\item{DT}{[ARGUMENT DEFINITION NEEDED]}

\item{strat_vars}{VARS TO STRATIFY THE DATA BY (PER REQUIREMENTS)}

\item{covariates}{[ARGUMENT DEFINITION NEEDED]}

\item{outcome_vars}{[ARGUMENT DEFINITION NEEDED]}
}
\description{
Use runPSM() to execute entire workflow. Alternatively utilize modular functions for
an interactive execution of this PSM analysis
}
\section{Functions}{
\itemize{
\item \code{caller_env}: Helper function for determining the parent caller (from package:rlang)

\item \code{runPSM}: Execute entire workflow

\item \code{get_result_psm}: Function that takes one hospital's data and appropriate argument to execute 
the matching (using internal function \code{getMatchedData}), then runs statistics across matched 
groups (using internal function \code{getStats})

\item \code{getMatchedData}: Takes in data from getDataQTI and returns a list of data.tables
with each list element being a PSM matched dataset containing

\item \code{getStats}: Function that takes the outcome of \code{getMatchedData} and 
runs statistical summary functions and performs a paired t-test across the psm matched
groups
}}

