% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataQTI.R
\name{QTI_data}
\alias{QTI_data}
\alias{getDataQTI}
\alias{getLevelsQTI}
\title{[TITLE NEEDED]}
\usage{
getDataQTI(hospID = NULL)

getLevelsQTI()
}
\arguments{
\item{hospID}{[DESCRIPTION OF ARGUMENT NEEDED]}
}
\value{
[DESCRIPTION OF OUTPUT NEEDED]
}
\description{
[DESCRIPTION OF FUNCTION NEEDED]
}
\section{Functions}{
\itemize{
\item \code{getDataQTI}: [DESCRIPTION OF FUNCTION NEEDED]

\item \code{getLevelsQTI}: [DESCRIPTION OF FUNCTION NEEDED]
}}

