% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_process_data.R
\name{query_data}
\alias{query_data}
\alias{getDataQTI}
\alias{getRawData}
\alias{split_check_hosp}
\title{Internal Data Query Function}
\usage{
getDataQTI(hospID = NULL, cols = "*")

getRawData()

split_check_hosp(DT, modal_A = "Open", modal_B = "Robotic")
}
\arguments{
\item{hospID}{[DESCRIPTION OF ARGUMENT NEEDED]}

\item{cols}{columns to limit query to for minor efficiency gains}

\item{DT}{Data for processing after query function is run}

\item{modal_A}{Modality 1 (defaults to Open)}

\item{modal_B}{Modality 2 (defaults to Robotic)}
}
\description{
Internal Data Query Function
}
\section{Functions}{
\itemize{
\item \code{getDataQTI}: Returns raw data with minor filtering on the server for efficiency

\item \code{getRawData}: Returns raw data with no filtering or cleaning

\item \code{split_check_hosp}: Function that splits a data set by hospital, and returns a list of hospitals that have enough 
patients to proceed with the PSM analysis
}}

