% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPSM.R
\name{runPSM}
\alias{runPSM}
\title{Run Full Analysis with PSM Matching}
\usage{
runPSM(hosp_id = 10112)
}
\arguments{
\item{hosp_id}{A numeric variable representing the hospital id}
}
\value{
[DOCUMENT RETURN VALUE]
}
\description{
Run Full Analysis with PSM Matching
}
