% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stat_data.R
\name{Stat_Data}
\alias{Stat_Data}
\alias{getStatData}
\title{Run Stats Across Groups}
\usage{
getStatData(matchDT, outcomes, groupby)
}
\arguments{
\item{matchDT}{Data that contains a column representing the outcome variables and contains columns CID and Modality for grouping}

\item{outcomes}{[ARGUMENT DEFINITION NEEDED]}

\item{groupby}{Column variables to group the data before running our desired statistics}
}
\description{
Run Stats Across Groups
}
\section{Functions}{
\itemize{
\item \code{getStatData}: Function that takes the outcome of \code{getMatchedData} and 
runs statistical summary functions and performs a paired t-test across the psm matched
groups
}}

