% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_results.R
\name{transformResult}
\alias{transformResult}
\title{Transform Final Output}
\usage{
transformResult(outDT)
}
\arguments{
\item{outDT}{Final step required due to requirements of how results should be structured}
}
\description{
Function to transform final result of runPSM per requirements. Note: current data layout is very sub-optimal.
}
