% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataQTI.R
\name{getDataQTI}
\alias{getDataQTI}
\alias{get_raw}
\title{[TITLE NEEDED]}
\usage{
getDataQTI(hosp_id = NULL)

get_raw(hosp_id = NULL)
}
\arguments{
\item{hosp_id}{[DESCRIPTION OF ARGUMENT NEEDED]}
}
\value{
[DESCRIPTION OF OUTPUT NEEDED]
}
\description{
[DESCRIPTION OF FUNCTION NEEDED]
}
\section{Functions}{
\itemize{
\item \code{get_raw}: helper function to perform query
}}

