% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatPSM.R
\name{getStatPSM}
\alias{getStatPSM}
\title{Test Difference of Means with Balancing}
\usage{
getStatPSM(DT, treatment, split_var, covars, outcome)
}
\arguments{
\item{DT}{Data set that meets IS specifications}

\item{treatment}{Name of the column that represents assignment of treatment or control group}

\item{split_var}{Variables use to stratify observations into groups}

\item{covars}{A character vector naming the columns that represent the covariates}

\item{outcome}{The name of the column that represents the outcome variable}
}
\value{
A data.table of statistics
}
\description{
Utilizes PSM to balance the groups. Model specified with the \code{treatment}, \code{covars},
and \code{outcome} arguments. The \code{split_var} argument is used for stratifying the data
prior to PSM matching.
}
