% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_data.R
\name{Match_Data}
\alias{Match_Data}
\alias{getMatchedData}
\title{Match Data Using PSM}
\usage{
getMatchedData(DT = NULL, strat_vars = NULL, covariates = NULL,
  outcome_vars = NULL)
}
\arguments{
\item{DT}{[ARGUMENT DEFINITION NEEDED]}

\item{strat_vars}{VARS TO STRATIFY THE DATA BY (PER REQUIREMENTS)}

\item{covariates}{[ARGUMENT DEFINITION NEEDED]}

\item{outcome_vars}{[ARGUMENT DEFINITION NEEDED]}
}
\description{
Match Data Using PSM
}
\section{Functions}{
\itemize{
\item \code{getMatchedData}: Takes in data from getDataQTI and returns a list of data.tables
with each list element being a PSM matched dataset containing
}}

