% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm_functions.R
\name{PSM_Functions}
\alias{PSM_Functions}
\alias{runPSM}
\alias{getMatchedData}
\alias{getStats}
\title{Run Full Analysis with PSM Matching}
\usage{
runPSM(hospID = NULL)

getMatchedData(DT = NULL, strat_vars = NULL, covariates = NULL,
  outcome_vars = NULL)

getStats(DT, outcome_vars)
}
\arguments{
\item{hospID}{A numeric variable representing the hospital id}

\item{DT}{[ARGUMENT DEFINITION NEEDED]}

\item{strat_vars}{VARS TO STRATIFY THE DATA BY (PER REQUIREMENTS)}

\item{covariates}{[ARGUMENT DEFINITION NEEDED]}

\item{outcome_vars}{[ARGUMENT DEFINITION NEEDED]}
}
\description{
Use runPSM() to execute entire workflow. Alternatively utilize modular functions for
an interactive execution of this PSM analysis
}
\section{Functions}{
\itemize{
\item \code{runPSM}: Execute entire workflow

\item \code{getMatchedData}: Takes in data from getDataQTI and returns a list of data.tables
with each list element being a PSM matched dataset containing

\item \code{getStats}: Function that takes the outcome of \code{getMatchedData} and 
runs statistical summary functions and performs a paired t-test across the psm matched
groups
}}

