% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_psm_functions.R
\name{PSM_Functions}
\alias{PSM_Functions}
\alias{caller_env}
\alias{runPSM}
\alias{run_match_stats}
\title{Run Full Analysis with PSM Matching}
\usage{
caller_env(n = 1)

runPSM(hospID = NULL, verbose = TRUE, wide = TRUE)

run_match_stats(hospDT, idcols, stratby, covars, outcomes, match_var)
}
\arguments{
\item{n}{internal argument for helper \code{caller_env}}

\item{hospID}{A numeric variable representing the hospital id}

\item{verbose}{A boolean indicating whether to print information about the connection and query being executed}

\item{wide}{A boolean indicating whether the data should be tranformed (default is TRUE per requirements)}

\item{hospDT}{Hospital data}

\item{idcols}{Columns to keep as IDs}

\item{stratby}{Columns to stratify the hospital data before PSM}

\item{covars}{Covariates to use to calculate the propensity score to match on}

\item{outcomes}{Outcome column variables to run statistics for across groups matched with PSM}
}
\description{
Use runPSM() to execute entire workflow. Alternatively utilize modular functions for
an interactive execution of this PSM analysis
}
\section{Functions}{
\itemize{
\item \code{caller_env}: Helper function for determining the parent caller (from package:rlang)

\item \code{runPSM}: Execute entire workflow

\item \code{run_match_stats}: Function that takes one hospital's data and appropriate argument to execute 
the matching (using internal function \code{getMatchedData}), then runs statistics across matched 
groups (using internal function \code{getStats})
}}

