% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPSM.R
\name{runPSM}
\alias{runPSM}
\title{Run Full Analysis with PSM Matching}
\usage{
runPSM(hospID = NULL, modal = "Open")
}
\arguments{
\item{hospID}{A numeric variable representing the hospital id}
}
\value{
[DOCUMENT RETURN VALUE]
}
\description{
Run Full Analysis with PSM Matching
}
